backend/requirements.txt
# ============================================================
# Zel-EYE: OI — FastAPI Backend Requirements
# Streamlit and its dependencies are intentionally excluded.
# ============================================================

# --- FastAPI core ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# --- Authentication ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# --- Database ---
duckdb>=0.10.0
supabase==2.15.0
boto3>=1.34.0

# --- Data processing ---
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# --- PTW PDF pipeline (ported unchanged from Streamlit) ---
python-docx>=1.1.0
PyPDF2>=3.0.0
reportlab>=4.0.0
Pillow>=10.0.0

# --- Utilities ---
requests>=2.31.0
tomli>=2.0.1; python_version < "3.11"
tqdm>=4.66.0
